requirements.txt
numpy==1.21.5
opencv_python==4.10.0.84
pdf2image==1.17.0
Pillow==9.0.1
Pillow==10.4.0
pytesseract==0.3.13
tqdm==4.66.5
